Allow to create a non G1 continuous shape
There was a test in the creation of the `Shape` to disallow the creation of
a Shape that is not G1 continuous. This blocked the use of the `Shape` object
for other use case.

This change allows to create non G1 continuous shapes but keeps that check
when it is needed in the algorithms.